volunteer upcoming deadline post show by useing mongodb  and slider create

+++ src/pages/ManagementPost.jsx
@@ -26,43 +26,62 @@ const ManagementPost = () => {
 
   //Delete
   const handleDelete = async (id) => {
-    try {
-      const { data } = await axios.delete(
-        `${import.meta.env.VITE_URL_SERVER}/delete/${id}`
-      );
-      console.log(data);
-      if (data.deletedCount > 0) {
-        Swal.fire({
-          title: "Deleted!",
-          text: "Your Product has been deleted.",
-          icon: "success",
-        });
-        const remaing = posts.filter((post) => post._id !== id);
-        setPosts(remaing);
+    const result = await Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+      cancelButtonText: "Cancel",
+    });
+
+    if (result.isConfirmed) {
+      try {
+        const { data } = await axios.delete(
+          `${import.meta.env.VITE_URL_SERVER}/delete/${id}`
+        );
+        if (data.deletedCount > 0) {
+          Swal.fire({
+            title: "Deleted!",
+            text: "Your Post has been deleted.",
+            icon: "success",
+          });
+          const remaining = posts.filter((post) => post._id !== id);
+          setPosts(remaining);
+        }
+      } catch (error) {
+        console.error("Error deleting post:", error);
       }
-    } catch (error) {
-      console.error("Error deleting product:", error);
     }
   };
 
+  if (!posts || posts.length === 0) {
+    return (
+      <div className="min-h-[calc(100vh-230px)] flex flex-col justify-center items-center gap-4">
+        <h2 className=" text-5xl font-semibold">
+          <i>NO DATA AVAILABLE</i>
+        </h2>
+        <p className=" text-3xl font-semibold">
+          <i>PLEASE! ADD SOME POST</i>
+        </p>
+        <Link
+          to="/addPost"
+          className="text-2xl font-medium btn text-white bg-gradient-to-r from-sky-500 to-indigo-500"
+        >
+          Add Post
+        </Link>
+      </div>
+    );
+  }
   if (loader) {
     <div className="min-h-[calc(100vh-230px)] flex justify-center items-center">
       <span className="loading loading-spinner loading-lg"></span>
     </div>;
   }
-
-  if (!posts) {
-    return (
-      <div className="">
-        <h2>NO DATA AVAILABLE</h2>
-        <p>PLEASE! ADD SOME POST</p>
-        <Link to="/addPost">Add Post</Link>
-      </div>
-    );
-  }
   return (
     <div>
-      <p>this is Management post</p>
       {loader ? (
         <div className="min-h-[calc(100vh-230px)] flex justify-center items-center">
           <span className="loading loading-spinner loading-lg"></span>
